Restructure symfony application part 2
