+++ htdocs/app/AppKernel.php
@@ -38,8 +38,7 @@ class AppKernel extends Kernel
             new Doctrine\Bundle\DoctrineBundle\DoctrineBundle(),
             new Sensio\Bundle\FrameworkExtraBundle\SensioFrameworkExtraBundle(),
             new Doctrine\Bundle\MigrationsBundle\DoctrineMigrationsBundle(),
-            new JMS\TranslationBundle\JMSTranslationBundle(),
-            new AppBundle\AppBundle(),
+            new JMS\TranslationBundle\JMSTranslationBundle()
         ];
 
         if (in_array($this->getEnvironment(), ['dev', 'test'], true)) {
