--- htdocs/src/AppBundle/AppBundle.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace AppBundle;
-
-use Symfony\Component\HttpKernel\Bundle\Bundle;
-
-class AppBundle extends Bundle
-{
-}
