--- htdocs/src/AppBundle/Exception/WrongDateFormatException.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace AppBundle\Exception;
-
-class WrongDateFormatException extends \Exception
-{
-}
