--- htdocs/src/AppBundle/Exception/WrongFileFormatException.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace AppBundle\Exception;
-
-class WrongFileFormatException extends \Exception
-{
-}
