--- htdocs/src/AppBundle/Repository/FieldNoteRepository.php
@@ -1,13 +1,0 @@
-<?php
-
-namespace AppBundle\Repository;
-
-/**
- * FieldNoteRepository
- *
- * This class was generated by the Doctrine ORM. Add your own custom
- * repository methods below.
- */
-class FieldNoteRepository extends \Doctrine\ORM\EntityRepository
-{
-}
