--- htdocs/src/AppBundle/Service/Interfaces/ErrorInterface.php
@@ -1,16 +1,0 @@
-<?php
-
-namespace AppBundle\Service\Interfaces;
-
-interface ErrorInterface
-{
-    /**
-     * @return bool
-     */
-    public function hasErrors();
-
-    /**
-     * @return array
-     */
-    public function getErrors();
-}
