+++ htdocs/src/Oc/FieldNotes/Entity/FieldNoteRepository.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Oc\FieldNotes\Entity;
+
+use Doctrine\ORM\EntityRepository;
+
+/**
+ * FieldNoteRepository
+ *
+ * This class was generated by the Doctrine ORM. Add your own custom
+ * repository methods below.
+ */
+class FieldNoteRepository extends EntityRepository
+{
+}
