+++ htdocs/src/Oc/FieldNotes/Exception/WrongDateFormatException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Oc\FieldNotes\Exception;
+
+use Exception;
+
+class WrongDateFormatException extends Exception
+{
+}
