+++ htdocs/src/Oc/FieldNotes/Exception/WrongFileFormatException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Oc\FieldNotes\Exception;
+
+use Exception;
+
+class WrongFileFormatException extends Exception
+{
+}
