+++ htdocs/src/Oc/Util/Error/ErrorInterface.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Oc\Util\Error;
+
+interface ErrorInterface
+{
+    /**
+     * @return bool
+     */
+    public function hasErrors();
+
+    /**
+     * @return array
+     */
+    public function getErrors();
+}
