+++ htdocs/src/OcLegacy/Template/LegacyTemplateTrait.php
@@ -1,0 +1,36 @@
+<?php
+
+namespace OcLegacy\Template;
+
+trait LegacyTemplateTrait
+{
+    /**
+     * @param int $menu
+     *
+     * @return void
+     */
+    protected function setMenu($menu)
+    {
+        $GLOBALS['tpl']->menuitem = $menu;
+    }
+
+    /**
+     * @param string $title
+     *
+     * @return void
+     */
+    protected function setTitle($title)
+    {
+        $GLOBALS['tpl']->title = $title;
+    }
+
+    /**
+     * @param string $url
+     *
+     * @return void
+     */
+    protected function setTarget($url)
+    {
+        $GLOBALS['tpl']->target = $url;
+    }
+}
